requirements.txt
keras==2.4.2
tensorflow-gpu==2.4.1
torch==1.8.2
torchvision==0.9.2
torchtext==0.9.2
matplotlib
opencv-python
numpy
scipy